:art: Integrate '__main__' module to managing initial import statements and global variables
